+++ Cargo.toml
@@ -43,6 +43,7 @@ ethcore-logger = { path = "logger" }
 ethcore-migrations = { path = "ethcore/migrations" }
 ethcore-miner = { path = "miner" }
 ethcore-network = { path = "util/network" }
+ethcore-service = { path = "ethcore/service" }
 ethcore-stratum = { path = "stratum" }
 ethcore-transaction = { path = "ethcore/transaction" }
 ethereum-types = "0.2"
@@ -130,6 +131,5 @@ members = [
 	"evmbin",
 	"miner",
 	"transaction-pool",
-	"whisper",
-	"util/rlp_compress"
+	"whisper"
 ]
