+++ ethcore/service/src/lib.rs
@@ -1,0 +1,32 @@
+// Copyright 2015-2018 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+extern crate ansi_term;
+extern crate ethcore;
+extern crate ethcore_io as io;
+extern crate kvdb;
+extern crate kvdb_rocksdb;
+extern crate stop_guard;
+
+#[macro_use]
+extern crate log;
+
+#[cfg(test)]
+extern crate tempdir;
+
+mod service;
+
+pub use service::ClientService;
