+++ ethcore/src/lib.rs
@@ -147,7 +147,6 @@ pub mod header;
 pub mod machine;
 pub mod miner;
 pub mod pod_state;
-pub mod service;
 pub mod snapshot;
 pub mod spec;
 pub mod state;
