+++ parity/informant.rs
@@ -22,9 +22,11 @@ use std::sync::{Arc};
 use std::sync::atomic::{AtomicUsize, AtomicBool, Ordering as AtomicOrdering};
 use std::time::{Instant, Duration};
 
-use ethcore::client::{BlockId, BlockChainClient, ChainInfo, BlockInfo, BlockChainInfo, BlockQueueInfo, ChainNotify, ClientReport, Client};
+use ethcore::client::{
+	BlockId, BlockChainClient, ChainInfo, BlockInfo, BlockChainInfo,
+	BlockQueueInfo, ChainNotify, ClientReport, Client, ClientIoMessage
+};
 use ethcore::header::BlockNumber;
-use ethcore::service::ClientIoMessage;
 use ethcore::snapshot::{RestorationStatus, SnapshotService as SS};
 use ethcore::snapshot::service::Service as SnapshotService;
 use ethsync::{LightSyncProvider, LightSync, SyncProvider, ManageNetwork};
