+++ sync/src/tests/consensus.rs
@@ -18,8 +18,7 @@ use std::sync::Arc;
 use hash::keccak;
 use ethereum_types::{U256, Address};
 use io::{IoHandler, IoContext, IoChannel};
-use ethcore::client::{Client, ChainInfo};
-use ethcore::service::ClientIoMessage;
+use ethcore::client::{Client, ChainInfo, ClientIoMessage};
 use ethcore::spec::Spec;
 use ethcore::miner::MinerService;
 use ethcore::account_provider::AccountProvider;
